requirements.txt
aiohttp >= 3.4.4
git+https://github.com/ebraminio/aiosseclient
multiprocessing-logging >= 0.2.6
py-dateutil >= 2.2
pylint
requests >= 2.20.1
schedule >= 0.5.0
nose >= 1.3.7

